Add camera graph and message synchronization (#35)
This patch adds the functionality to build a camera graph. An undirected
graph, which nodes are the cameras and the edges the amount of shared
observations across the whole dataset.

For finding the shared observations we need to match the images taken from
multiple cameras at the same timestamp, so this patch includes an iterator
to approximately synchronize messages across all cameras.